Pricing change and extracted oracle logic
Now the pricing for each policy depending on the price drop includes not
only the lossProb but also the junior and total collateralization
ratios.

Also, I extracted all the oracle related logic to an external contract
(IPriceOracle).
This contract has a simple interface that just returns the currentPrice
in wad.

The chainlink related code is now moved to ChainlinkPriceOracle that
implements the IPriceOracle interface using two underlying chainlink
oracles (asset and reference).